requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0

jinja2==3.1.6
aiofiles==25.1.0

python-multipart==0.0.22

itsdangerous==2.2.0

starlette==0.52.1

pydantic==2.12.5

python-jose[cryptography]==3.3.0

passlib[bcrypt]==1.7.4
bcrypt==4.1.3

python-dotenv==1.2.2
